Simplify Annotation: don't fire events
Using the DOM events to coordinate between the collection object
(Annotations) and Annotation instances is problematic. We need to be
able to add tracker expandos to the nodes associated with annotation
targets before any external event listeners are called.

So instead of having Annotations fire events, move this responsibility
to the collection object.