Add getAnnotations method for querying annotations
This commit adds support for querying an Annotations collection with a
list of DOM nodes, in order to find annotations that reference those
nodes.

We use an expando technique (similar to that used by jQuery.data) in
order to link from DOM nodes to annotations without holding references
to nodes in the Annotations collection object.

There is more to do here, because at the moment an Annotation object
still holds references to nodes (through its targets).